+++ packages/artifact/src/internal/requestUtils.ts
@@ -1,4 +1,4 @@
-import {IHttpClientResponse} from '@actions/http-client/interfaces'
+import {IHttpClientResponse} from '@actions/http-client/lib/interfaces'
 import {
   isRetryableStatusCode,
   isSuccessStatusCode,