+++ packages/cache/src/internal/contracts.d.ts
@@ -1,5 +1,5 @@
 import {CompressionMethod} from './constants'
-import {ITypedResponse} from '@actions/http-client/interfaces'
+import {ITypedResponse} from '@actions/http-client/lib/interfaces'
 import {HttpClientError} from '@actions/http-client'
 
 export interface ITypedResponseWithError<T> extends ITypedResponse<T> {